refactor pinnepulickeys
[changelog]
internal: refactor pinnedpublickeys.
